Added Global Path Classification for constraint spaces detection

+++ nav2_route/src/node_spatial_tree.cpp
@@ -61,7 +61,7 @@ bool NodeSpatialTree::findNearestGraphNodesToPose(
   const geometry_msgs::msg::PoseStamped & pose_in, std::vector<unsigned int> & node_ids)
 {
   size_t num_results = static_cast<size_t>(num_of_nearest_nodes_);
-  std::vector<unsigned int> ret_index(num_results);
+  std::vector<size_t> ret_index(num_results);
   std::vector<double> out_dist_sqr(num_results);
   const double query_pt[2] = {pose_in.pose.position.x, pose_in.pose.position.y};
   num_results = kdtree_->knnSearch(&query_pt[0], num_results, &ret_index[0], &out_dist_sqr[0]);
